Add files via upload
The purpose of this project was to develop a K-means clustering algorithm to group customers of a retail store based on their purchase history. Clustering customers can help in understanding different customer segments, enabling targeted marketing strategies and improved customer service.

The dataset used in this project which was collected which was downloaded from kaggle contains the purchase history of customers. Each row represents a customer, and the columns include features such as customer ID, gender, age and annual income and their spending score. 

The data was then preprocessed by dropping irrelevant features such as the customer ID, age and the gender.

Data Preprocessing: Before applying the K-means clustering algorithm, I needed to preprocess the data to ensure it was suitable for analysis. Here’s what I did:

Loading Data: I loaded the dataset into a pandas DataFrame.

Handling Missing Values: I identified and handled any missing values in the dataset. Missing values were either filled with suitable statistics (e.g., mean, median) or removed if necessary.

Feature Scaling: The features to ensure they were on a similar scale, which is important for the K-means algorithm. StandardScaler from sklearn was used to standardize the features.

Determining the Number of Clusters: Choosing the optimal number of clusters (K) is crucial for effective clustering. Elbow Method was adopted to determine the appropriate value of K. This method involves plotting the inertia (sum of squared distances from each point to its assigned cluster center) for different values of K and identifying the point where the rate of decrease sharply slows down (the "elbow").


Training the K-means Model: Based on the Elbow Method, I selected the optimal number of clusters. Then, the K-means model was initialized and trained with this value of K.

Visualizing the Clusters: To help understand how customers are grouped, I visualized the clusters and created scatter plots to visualize the clustering in a two-dimensional space, which helped in interpreting the clusters and understanding the characteristics of each group.
Interpreting Results.
Finally, the results of the clustering were interpreted:

Cluster Characteristics: Each cluster was analyzed to understand the characteristics of the customers within it. For example, one cluster represents high- annual income, high spending score, while another represents low annual income, low spending score and average parameters respectively.

Marketing Strategies: Based on the characteristics of each cluster, targeted marketing strategies can be implemented. For instance, customers with high annual income could be targeted with loyalty programs, while occasional buyers might be offered special discounts to increase their spending score.